requirement.txt
ipykernel==7.1.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
matplotlib-inline==0.2.1
numpy==2.3.4
pandas==2.3.3
pip==25.2
psycopg2==2.9.11
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.5
SQLAlchemy==2.0.44
urllib3==2.5.0

